Fixed warnings when compiling segmentation.

+++ src/segmentation/aploc.c
@@ -14,23 +14,23 @@
 
 # define foreach_state(_state_)     \
     for (int i = 0 ; i < 5 ; i++)
-        
+
 # define PRECISION 1
 
 # include "aploc.h"
 
-static inline                                                                   
-int get_BW(SDL_Surface *img, int x, int y) //returns 0 if black, 1 if white     
-{                                                                               
-     Uint32 pixel = getpixel(img, x, y);                                         
-     Uint8 r, g, b;                                                              
+static inline
+int get_BW(SDL_Surface *img, int x, int y) //returns 0 if black, 1 if white
+{
+     Uint32 pixel = getpixel(img, x, y);
+     Uint8 r, g, b;
      SDL_GetRGB(pixel, img->format, &r, &g, &b);
      //printf("color at %d %d: %d \n", x, y, g);
-     //putpixel(img, x, y, SDL_MapRGB(img->format, 255, 0, 0)); 
-     if(g == 0)                                                                  
-        return 0;                                                               
-     else                                                                        
-        return 1; 
+     //putpixel(img, x, y, SDL_MapRGB(img->format, 255, 0, 0));
+     if(g == 0)
+        return 0;
+     else
+        return 1;
 }
 
 static inline
@@ -39,17 +39,17 @@ int check_ratio(int *state)
     int totsize = 0;
     for(int i = 1; i < 4; i++)
         totsize += state[i];
-    
+
     if (totsize < 3)
         return 0;
-   
+
     //printf("state %d %d %d | %lf\n", state[0], state[1], state[2], msize);
 
     double msize = totsize/3;
     double max_var = msize/1.5;
-    if((abs(msize - state[1]) <= max_var) &&
-       (abs(msize - state[2]) <= max_var) &&
-       (abs(msize - state[3]) <= max_var))
+    if((abs((int)msize - state[1]) <= max_var) &&
+       (abs((int)msize - state[2]) <= max_var) &&
+       (abs((int)msize - state[3]) <= max_var))
         return totsize;
     else
         return 0;
@@ -69,9 +69,9 @@ double check_ver(SDL_Surface *img, int start_y, int center_x, int totalsize)
     int h = img->h;
     int state_check[5] = {0};
     int y = start_y;
-    
+
     //upwards check from center
- 
+
     while((y >= 0) && (get_BW(img, center_x, y) == 0))
     {
         state_check[2]++ ;
@@ -95,10 +95,10 @@ double check_ver(SDL_Surface *img, int start_y, int center_x, int totalsize)
     }
     if(y < 0)
         return 0;
-    
+
     //downwards check from center
-    
-    y = start_y + 1; 
+
+    y = start_y + 1;
 
     while((y <= h) && (get_BW(img, center_x, y) == 0))
     {
@@ -129,7 +129,7 @@ double check_ver(SDL_Surface *img, int start_y, int center_x, int totalsize)
     {
         check_totalsize += state_check[i];
     }
-    check_totalsize -= state_check[4]; 
+    check_totalsize -= state_check[4];
     check_totalsize -= state_check[0];
     if(5 * abs(check_totalsize - totalsize) >= 2 * totalsize)
         return 0;
@@ -137,7 +137,7 @@ double check_ver(SDL_Surface *img, int start_y, int center_x, int totalsize)
     double center = get_center(check_totalsize, y - state_check[4]);
     if(check_ratio(state_check) != 0)
         return center;
-    return 0;    
+    return 0;
 }
 
 static inline
@@ -146,9 +146,9 @@ double check_hor(SDL_Surface *img, int center_y, int start_x, int totalsize)
     int w = img->w;
     int state_check[5] = {0};
     int x = start_x;
-    
+
     //leftwards check from center
-    
+
     while((x >= 0) && (get_BW(img, x, center_y) == 0))
     {
         state_check[2]++ ;
@@ -156,7 +156,7 @@ double check_hor(SDL_Surface *img, int center_y, int start_x, int totalsize)
     }
     if(x < 0)
         return 0;
- 
+
     while((x >= 0) && (get_BW(img, x, center_y) == 1))
     {
         state_check[1]++ ;
@@ -172,11 +172,11 @@ double check_hor(SDL_Surface *img, int center_y, int start_x, int totalsize)
     }
     if(x < 0)
         return 0;
-    
+
     //rightwards check from center
-    
-    x = start_x + 1; 
-    
+
+    x = start_x + 1;
+
     while((x <= w) && (get_BW(img, x, center_y) == 0))
     {
         state_check[2]++ ;
@@ -184,7 +184,7 @@ double check_hor(SDL_Surface *img, int center_y, int start_x, int totalsize)
     }
     if(x > w)
         return 0;
- 
+
     while((x <= w) && (get_BW(img, x, center_y) == 1))
     {
         state_check[3]++ ;
@@ -200,7 +200,7 @@ double check_hor(SDL_Surface *img, int center_y, int start_x, int totalsize)
     }
     if(x > w)
         return 0;
-    
+
     int check_totalsize = 0;
     foreach_state (state_check)
     {
@@ -210,22 +210,22 @@ double check_hor(SDL_Surface *img, int center_y, int start_x, int totalsize)
     check_totalsize -= state_check[0];
     if(5 * abs(check_totalsize - totalsize) >= 2 * totalsize)
         return 0;
-    
+
     double center = get_center(check_totalsize, x - state_check[4]);
     if(check_ratio(state_check) != 0)
         return center;
-    return 0;    
+    return 0;
 }
 
 static inline
 double check_diag(SDL_Surface *img, int center_y, int center_x, int totalsize)
 {
     int state_check[5] = {0};
-    
+
     //to top-left check from center
-    
+
     int i = 0;
-    
+
     while((center_y >= i) && (center_x >= i) && (get_BW(img, center_x - i, center_y - i) == 0))
     {
         state_check[2]++ ;
@@ -233,7 +233,7 @@ double check_diag(SDL_Surface *img, int center_y, int center_x, int totalsize)
     }
     if(center_y < i || center_x < i)
         return 0;
- 
+
     while((center_y >= i) && (center_x >= i) && (get_BW(img, center_x - i, center_y - i) == 1))
     {
         state_check[1]++ ;
@@ -247,13 +247,13 @@ double check_diag(SDL_Surface *img, int center_y, int center_x, int totalsize)
         state_check[0]++ ;
         i++;
     }
-    
+
     //to bottom-right check from center
     int h = img->h;
     int w = img->w;
-    
-    i = 1; 
-    
+
+    i = 1;
+
     while(((center_y + i) < h) && (center_x+1) < w && (get_BW(img, center_x + i, center_y + i) == 0))
     {
         state_check[2]++ ;
@@ -261,7 +261,7 @@ double check_diag(SDL_Surface *img, int center_y, int center_x, int totalsize)
     }
     if((center_y + i) > h || (center_x + i) >= w)
         return 0;
- 
+
     while(((center_y + i) < h) && ((center_x + i) < w) && (get_BW(img, center_x + i, center_y + i) == 1))
     {
         state_check[3]++ ;
@@ -290,7 +290,7 @@ double check_diag(SDL_Surface *img, int center_y, int center_x, int totalsize)
 
     if(check_ratio(state_check) != 0)
         return 1;
-    return 0;    
+    return 0;
 }
 
 static inline
@@ -302,7 +302,7 @@ int handle_centers(SDL_Surface *img, struct Dmat *centers, int *state, int y,
     double center_y = check_ver(img, y, (int)center_x, totalsize);
     if(center_y == 0)
         return 0;
-    
+
     center_x = check_hor(img, center_y, center_x, totalsize);
     if(center_x == 0)
         return 0;
@@ -310,25 +310,25 @@ int handle_centers(SDL_Surface *img, struct Dmat *centers, int *state, int y,
     int validation = check_diag(img, center_y, center_x, totalsize);
     if(validation == 0)
         return 0;
-    
+
     double *new_center = calloc(2, sizeof(double));
     new_center[0] = center_x;
     new_center[1] = center_y;
     int found = 0;
-    
+
     foreach_line(centers)
     {
-        double dist_x = abs(mat[i][0] - new_center[0]);
-        double dist_y = abs(mat[i][1] - new_center[1]);
+        double dist_x = abs((int)mat[i][0] - (int)new_center[0]);
+        double dist_y = abs((int)mat[i][1] - (int)new_center[1]);
         if(dist_x < 10 && dist_y < 10)
         {
-            mat[i][0] = (mat[i][0] + new_center[0]) / 2.0; 
-            mat[i][1] = (mat[i][1] + new_center[1]) / 2.0;
+            mat[i][0] = (mat[i][0] + (int)new_center[0]) / 2.0;
+            mat[i][1] = (mat[i][1] + (int)new_center[1]) / 2.0;
             found = 1;
             break;
         }
     }
-    
+
     if(found == 0)
     {
         add_Dmat(centers, new_center);
@@ -345,11 +345,11 @@ void ScanAP(SDL_Surface *img, double *Px, double *Py)
     int TLy = *Py - scanS/2;
     int BRx = *Px + scanS/2;
     int BRy = *Py + scanS/2;
-    
+
     struct Dmat *centers = init_Dmat(2, 2);
     int *state = calloc(5, sizeof(int));
     int state_count = 0;
-    
+
     //Scan
     for(int y = TLy; y < BRy; y++)
     {
@@ -386,7 +386,7 @@ void ScanAP(SDL_Surface *img, double *Px, double *Py)
                             state[4] = 0;
                             continue;
                         }
-                        
+
                         RESET_STATE(state);
                         state_count = 0;
                     }
@@ -399,12 +399,12 @@ void ScanAP(SDL_Surface *img, double *Px, double *Py)
             }
         }
     }
-    
-    free(state);   
+
+    free(state);
     foreach_line(centers)
     {
-        double dist_x = abs(mat[i][0] - *Px)/ 2.0;
-        double dist_y = abs(mat[i][1] - *Py)/ 2.0;
+        double dist_x = abs((int)mat[i][0] - (int)*Px)/ 2.0;
+        double dist_y = abs((int)mat[i][1] - (int)*Py)/ 2.0;
         if(dist_x < 10 && dist_y < 10)
         {
             *Px = mat[i][0];
